Move dot navigation + add dedicated scroll func
