+++ src/components/dot-navigation.js
@@ -1,5 +1,5 @@
-import { cx } from "../../utils/cx";
-import { DotIcon } from "../../components/dot-icon";
+import { cx } from "../utils/cx";
+import { DotIcon } from "./dot-icon";
 
 export function DotNavigation({ className, items, activeItem, onNavigate }) {
   return (
@@ -19,4 +19,12 @@ export function DotNavigation({ className, items, activeItem, onNavigate }) {
       ))}
     </div>
   );
+}
+
+export function scrollIntoView(element) {
+  element.scrollIntoView({
+    behavior: "smooth",
+    block: "nearest",
+    inline: "center",
+  });
 }
