+++ src/screens/resources/websites-mobile.js
@@ -1,21 +1,15 @@
 import { websites } from "./resources";
-import { DotNavigation } from "../videos/dot-navigation";
+import { DotNavigation, scrollIntoView } from "../../components/dot-navigation";
 import { cx } from "../../utils/cx";
-import { videos } from "../videos/videos";
 import { useRef, useState } from "react";
 
 export function WebsitesMobile({ className }) {
-  const [visibleWebsiteKey, setVisibleWebsiteKey] = useState(videos[0].id);
+  const [visibleWebsiteKey, setVisibleWebsiteKey] = useState(websites[0].key);
   const websitesRef = useRef({});
 
   const scrollTo = (websiteKey) => {
     setVisibleWebsiteKey(websiteKey);
-
-    websitesRef.current[websiteKey].scrollIntoView({
-      behavior: "smooth",
-      block: "nearest",
-      inline: "center",
-    });
+    scrollIntoView(websitesRef.current[websiteKey]);
   };
 
   return (
@@ -45,7 +39,7 @@ export function WebsitesMobile({ className }) {
             <p>{website.description}</p>
 
             <a
-              className="ml-auto mt-auto rounded-full bg-brown px-5 py-1 font-bold uppercase text-white"
+              className="bg-brown ml-auto mt-auto rounded-full px-5 py-1 font-bold uppercase text-white"
               href={website.url}
               target="_blank"
               rel="noreferrer"
